requirements.txt
altair==5.3.0
annotated-types==0.6.0
anthropic==0.29.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.1
executing==2.0.1
feedparser==6.0.11
filelock==3.14.0
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
groq==0.4.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hugchat==0.4.1
huggingface-hub==0.23.1
idna==3.6
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
ollama==0.1.8
openai==1.16.2
packaging==24.0
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.45
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.18.0
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.34.0
regex==2024.5.15
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
streamlit==1.35.0
style==1.1.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.41.1
typing_extensions==4.12.0
tzdata==2024.1
update==0.0.1
urllib3==2.2.1
validators==0.28.1
watchdog==4.0.0
wcwidth==0.2.13
zipp==3.19.0
